Add support for copy_ttl_out action
This adds support for the OpenFlow 1.1+ copy_ttl_out action.
And also adds an NX copy_ttl_out action.

The implementation does not support copying in the case where the outermost
header is IP as it is unclear to me that Open vSwtich has a notion of an
inner IP header to copy the TLL from.

The handling of the TTL modification is entirely handled in userspace.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

This patch depends on "flow: Set ttl in flow_compose()."